Finalized Little Hexy model
