+++ data/copters/prop.obj
@@ -1,0 +1,264 @@
+####
+#
+# OBJ File Generated by Meshlab
+#
+####
+# Object prop.obj
+#
+# Vertices: 63
+# Faces: 122
+#
+####
+vn 0.095645 -0.994705 -0.037628
+v 0.027377 -0.013215 0.002600
+vn -0.088337 -0.995289 -0.039961
+v -0.000674 -0.010716 0.002351
+vn -0.088332 -0.995288 -0.040003
+v -0.000613 -0.010523 -0.002574
+vn 0.095642 -0.994704 -0.037648
+v 0.027438 -0.013023 -0.002327
+vn 0.324443 0.944932 -0.042918
+v 0.030451 0.007902 -0.001473
+vn 0.089108 0.994536 -0.054391
+v 0.001253 0.010504 -0.001730
+vn 0.086789 0.987172 0.134013
+v 0.001192 0.010312 0.003196
+vn 0.322617 0.938326 0.124352
+v 0.030390 0.007710 0.003454
+vn -0.179990 0.981329 -0.067814
+v -0.027946 0.013106 -0.001988
+vn -0.813993 0.580839 -0.006441
+v -0.034160 0.010243 0.000006
+vn -0.151389 0.981227 0.119485
+v -0.028006 0.012914 0.002938
+vn -0.269322 -0.962183 -0.040862
+v -0.028724 -0.008216 0.002105
+vn -0.269384 -0.962165 -0.040878
+v -0.028665 -0.008024 -0.002822
+vn 0.761918 -0.647472 -0.016200
+v 0.034186 -0.011164 0.000008
+vn 0.916858 0.398568 0.022697
+v 0.036028 0.004005 0.000622
+vn -0.440538 -0.896820 -0.040506
+v -0.035057 -0.005101 0.002149
+vn -0.882877 -0.468049 -0.038202
+v -0.035494 -0.004766 -0.000526
+vn -0.675476 -0.440649 0.591237
+v -0.035075 -0.005210 0.004749
+vn -0.236811 -0.970790 -0.038565
+v 0.050750 0.006080 0.000885
+vn 0.079907 0.996183 0.035157
+v 0.048738 -0.016583 0.000046
+vn -0.310277 0.950093 0.032441
+v 0.075973 -0.036484 -0.000398
+vn 0.000000 0.000000 0.000000
+v 0.256910 -0.037364 0.001788
+vn -0.630849 -0.775042 -0.036627
+v 0.080154 0.009954 0.001397
+vn 0.210046 -0.000000 -0.977692
+v 0.258921 -0.013012 0.002675
+vn 0.173639 -0.984215 -0.034212
+v 0.065423 0.008017 0.001140
+vn 0.706870 0.706544 0.033629
+v 0.062629 -0.023450 -0.000052
+vn 0.024833 -0.031174 -0.999206
+v 0.275711 -0.037430 0.001944
+vn 0.000000 0.000000 0.000000
+v 0.276971 -0.023234 0.002514
+vn 0.046827 0.604133 0.795507
+v 0.001098 0.009571 0.006399
+vn 0.194464 0.621824 0.758630
+v 0.030161 0.006982 0.006655
+vn -0.017137 -0.041439 0.998994
+v 0.000912 0.007468 0.006314
+vn 0.675956 0.319615 0.664026
+v 0.035376 0.004013 0.006603
+vn 0.063583 -0.766391 0.639221
+v 0.027344 -0.013317 0.005211
+vn -0.075573 -0.787035 0.612263
+v -0.000707 -0.010819 0.004997
+vn 0.686614 0.368789 -0.626543
+v 0.035515 0.004457 -0.004778
+vn 0.006386 0.036650 -0.999308
+v 0.001051 0.007912 -0.005067
+vn 0.060088 0.662663 -0.746504
+v 0.001239 0.010016 -0.004983
+vn 0.212510 0.678901 -0.702804
+v 0.030302 0.007426 -0.004726
+vn -0.660463 0.439711 0.608641
+v -0.034221 0.010094 0.005356
+vn -0.117767 0.567162 0.815144
+v -0.027965 0.012161 0.006143
+vn -0.066157 -0.750962 -0.657024
+v -0.000582 -0.010421 -0.005185
+vn 0.085723 -0.728948 -0.679181
+v 0.027469 -0.012921 -0.004938
+vn -0.125483 0.624083 -0.771217
+v -0.027826 0.012606 -0.005239
+vn -0.639221 0.508807 -0.576639
+v -0.034097 0.010489 -0.004769
+vn -0.209350 -0.741616 0.637322
+v -0.028757 -0.008318 0.004716
+vn -0.192251 -0.688455 -0.699335
+v -0.028632 -0.007922 -0.005432
+vn -0.023807 -0.130653 0.991143
+v -0.000434 -0.007690 0.005707
+vn 0.605753 -0.551480 0.573528
+v 0.034116 -0.011374 0.005364
+vn 0.619099 -0.503127 -0.602976
+v 0.034236 -0.010979 -0.004768
+vn -0.001088 -0.068157 -0.997674
+v -0.000333 -0.007680 -0.005693
+vn -0.664798 -0.386251 -0.639418
+v -0.034951 -0.004814 -0.005379
+vn -0.021164 -0.038163 0.999048
+v 0.000239 -0.000111 0.006011
+vn 0.998026 0.061317 0.013567
+v -0.048353 0.016020 0.000057
+vn 0.991122 -0.132756 0.007288
+v -0.050166 -0.006703 -0.000780
+vn -0.313271 0.949090 0.033025
+v -0.075587 0.035868 0.000499
+vn -0.397724 0.917276 0.020529
+v -0.256284 0.036746 -0.001684
+vn -0.243069 -0.968572 -0.052788
+v -0.079511 -0.010577 -0.001293
+vn -0.150865 -0.522664 0.839084
+v -0.257432 0.011962 -0.002093
+vn -0.881026 -0.455066 -0.129262
+v -0.259366 0.011996 -0.002617
+vn 0.176129 -0.983785 -0.033869
+v -0.064839 -0.008640 -0.001036
+vn 0.712885 0.700486 0.033389
+v -0.062351 0.022940 0.000155
+vn 0.891024 -0.447214 0.077954
+v -0.275072 0.037430 -0.001888
+vn 0.000000 0.000000 0.000000
+v -0.276981 0.022908 -0.002408
+# 63 vertices, 0 vertices normals
+
+f 1//1 2//2 3//3
+f 3//3 4//4 1//1
+f 5//5 6//6 7//7
+f 7//7 8//8 5//5
+f 9//9 10//10 11//11
+f 3//3 2//2 12//12
+f 12//12 13//13 3//3
+f 7//7 6//6 9//9
+f 9//9 11//11 7//7
+f 1//1 4//4 14//14
+f 5//5 8//8 15//15
+f 12//12 16//16 13//13
+f 16//16 17//17 13//13
+f 17//17 16//16 18//18
+f 19//19 15//15 14//14
+f 14//14 20//20 19//19
+f 20//20 14//14 15//15
+f 15//15 19//19 20//20
+f 21//21 22//22 23//23
+f 23//23 22//22 24//24
+f 21//21 23//23 22//22
+f 23//23 24//24 22//22
+f 25//25 26//26 21//21
+f 21//21 23//23 25//25
+f 26//26 25//25 23//23
+f 23//23 21//21 26//26
+f 19//19 20//20 26//26
+f 26//26 25//25 19//19
+f 20//20 19//19 25//25
+f 25//25 26//26 20//20
+f 22//22 24//24 27//27
+f 24//24 28//28 27//27
+f 8//8 7//7 29//29
+f 29//29 30//30 8//8
+f 30//30 29//29 31//31
+f 31//31 32//32 30//30
+f 33//33 34//34 2//2
+f 2//2 1//1 33//33
+f 35//35 36//36 37//37
+f 37//37 38//38 35//35
+f 11//11 10//10 39//39
+f 39//39 40//40 11//11
+f 32//32 15//15 8//8
+f 8//8 30//30 32//32
+f 4//4 3//3 41//41
+f 41//41 42//42 4//4
+f 35//35 38//38 5//5
+f 5//5 15//15 35//35
+f 38//38 37//37 6//6
+f 6//6 5//5 38//38
+f 43//43 44//44 10//10
+f 10//10 9//9 43//43
+f 34//34 45//45 12//12
+f 12//12 2//2 34//34
+f 41//41 3//3 13//13
+f 13//13 46//46 41//41
+f 37//37 36//36 44//44
+f 44//44 43//43 37//37
+f 6//6 37//37 43//43
+f 43//43 9//9 6//6
+f 29//29 7//7 11//11
+f 11//11 40//40 29//29
+f 31//31 29//29 40//40
+f 40//40 39//39 31//31
+f 34//34 47//47 18//18
+f 18//18 45//45 34//34
+f 48//48 47//47 34//34
+f 34//34 33//33 48//48
+f 1//1 14//14 48//48
+f 48//48 33//33 1//1
+f 42//42 49//49 14//14
+f 14//14 4//4 42//42
+f 42//42 41//41 50//50
+f 50//50 49//49 42//42
+f 51//51 50//50 41//41
+f 41//41 46//46 51//51
+f 17//17 51//51 46//46
+f 46//46 13//13 17//17
+f 18//18 16//16 12//12
+f 12//12 45//45 18//18
+f 32//32 52//52 47//47
+f 32//32 47//47 48//48
+f 48//48 14//14 15//15
+f 15//15 32//32 48//48
+f 49//49 35//35 15//15
+f 15//15 14//14 49//49
+f 49//49 50//50 36//36
+f 36//36 35//35 49//49
+f 44//44 36//36 50//50
+f 50//50 51//51 44//44
+f 10//10 44//44 51//51
+f 51//51 17//17 10//10
+f 39//39 10//10 17//17
+f 17//17 18//18 39//39
+f 53//53 10//10 17//17
+f 17//17 54//54 53//53
+f 54//54 17//17 10//10
+f 10//10 53//53 54//54
+f 55//55 56//56 57//57
+f 57//57 56//56 58//58
+f 56//56 59//59 58//58
+f 55//55 57//57 56//56
+f 57//57 59//59 56//56
+f 60//60 61//61 55//55
+f 55//55 57//57 60//60
+f 61//61 60//60 57//57
+f 57//57 55//55 61//61
+f 54//54 53//53 61//61
+f 61//61 60//60 54//54
+f 53//53 54//54 60//60
+f 60//60 61//61 53//53
+f 56//56 59//59 62//62
+f 59//59 63//63 62//62
+f 52//52 39//39 47//47
+f 47//47 39//39 18//18
+f 22//22 27//27 24//24
+f 24//24 27//27 28//28
+f 56//56 62//62 59//59
+f 59//59 62//62 63//63
+f 58//58 59//59 57//57
+f 31//31 52//52 32//32
+f 31//31 39//39 52//52
+# 122 faces, 0 coords texture
+
+# End of File
